# openapi.yaml
openapi: 3.0.3
info:
  title: API MongoDB — Lista de Exercícios
  version: 1.0.0
  description: |-
    Lista de Exercícios de MongoDB implementada como API REST com **Node.js + Express**
    e o driver oficial `mongodb` (sem ODM).

    Cada grupo de rotas corresponde a uma parte da lista:

    | Recurso | Exercícios | Conceitos |
    |---|---|---|
    | Produtos | 2 a 11 | CRUD, `$gt`/`$lt`, `$and`/`$or`/`$nor`, projeção, `sort`/`limit`, `$set`, `$inc` |
    | Clientes | 12, 13, 14 | documentos aninhados, `$push`, `$pull`, `$addToSet` |
    | Filmes | 15 | modelagem com array + documento aninhado |
    | Alunos | 16 | arrays de valores, `$avg`, `$unwind` |
    | Vendas | 17, 18, 19 | `$match`, `$group`, `$project`, `$multiply`, `$sum`, `$avg` |
    | Pedidos | 20 | array de documentos, `$unwind`, agrupamento por status |
    | Exercícios | todos | enunciado + comando no mongosh + resultado do banco |

    A rota **`/api/exercicios`** devolve, para cada item da lista, o enunciado, o comando
    equivalente no `mongosh` e o resultado real vindo do banco.
  contact:
    name: Repositório do trabalho
    url: http://localhost:3000
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Ambiente local
...
components:
  schemas:
    ObjectId:
      type: string
      pattern: ^[0-9a-fA-F]{24}$
      description: Identificador único do documento (24 caracteres hexadecimais).
      example: 6a98c80f3e8030cb5731e645
...
    Filme:
      type: object
      required:
        - titulo
        - ano
      properties:
        _id:
          $ref: '#/components/schemas/ObjectId'
        titulo:
          type: string
          example: Interestelar
        ano:
          type: integer
          example: 2014
        classificacaoIndicativa:
          type: string
          example: 12 anos
        duracaoMinutos:
          type: integer
          example: 169
        generos:
          type: array
          items:
            type: string
          example:
            - Ficção científica
            - Aventura
            - Drama
        diretor:
          type: object
          properties:
            nome:
              type: string
              example: Christopher Nolan
            nacionalidade:
              type: string
              example: Britânico-americana
            anoNascimento:
              type: integer
              example: 1970
        atores:
          type: array
          items:
            type: object
            properties:
              nome:
                type: string
                example: Anne Hathaway
              personagem:
                type: string
                example: Dra. Brand
        avaliacao:
          type: object
          properties:
            media:
              type: number
              example: 8.7
            totalVotos:
              type: integer
              example: 1980000
            fonte:
              type: string
              example: IMDb
    ListaFiltrada:
      type: object
      description: 'Resposta padrão das consultas: repete o filtro aplicado e devolve os dados.'
      properties:
        filtro:
          type: object
          description: Filtro montado a partir da query string.
        total:
          type: integer
          example: 5
        dados:
          type: array
          items:
            type: object
...
    ResultadoRemocao:
      type: object
      properties:
        removidos:
          type: integer
          example: 1
...
    Erro:
      type: object
      properties:
        erro:
          type: string
          example: Recurso não encontrado.
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      description: ObjectId do documento.
      schema:
        $ref: '#/components/schemas/ObjectId'
...
  responses:
    RequisicaoInvalida:
      description: Dados inválidos ou campos obrigatórios ausentes.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Erro'
    NaoEncontrado:
      description: Documento não encontrado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Erro'
...
paths:
...
  /api/filmes:
    get:
      tags:
        - Filmes
      summary: Lista filmes (exercício 15)
      description: >
        Demonstra os três tipos de consulta da modelagem proposta: array de strings (`generos`), documento aninhado
        (`diretor.nome`) e array de documentos (`atores.nome`).
      parameters:
        - name: genero
          in: query
          schema:
            type: string
          example: Drama
          description: Valor dentro do array generos.
        - name: diretor
          in: query
          schema:
            type: string
          example: Christopher Nolan
          description: Campo diretor.nome.
        - name: ator
          in: query
          schema:
            type: string
          example: Anne Hathaway
          description: Campo atores.nome.
        - name: ano
          in: query
          schema:
            type: integer
          example: 2014
        - name: notaMin
          in: query
          schema:
            type: number
          example: 8
          description: Campo avaliacao.media com $gte.
      responses:
        '200':
          description: Lista de filmes.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ListaFiltrada'
                  - type: object
                    properties:
                      dados:
                        type: array
                        items:
                          $ref: '#/components/schemas/Filme'
    post:
      tags:
        - Filmes
      summary: Cadastra um filme (exercício 15)
      description: Documento com valores simples, arrays (generos, atores) e documentos aninhados (diretor, avaliacao).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Filme'
      responses:
        '201':
          description: Filme criado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Filme'
        '400':
          $ref: '#/components/responses/RequisicaoInvalida'
  /api/filmes/{id}:
    get:
      tags:
        - Filmes
      summary: Busca um filme pelo _id
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Filme encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Filme'
        '400':
          $ref: '#/components/responses/RequisicaoInvalida'
        '404':
          $ref: '#/components/responses/NaoEncontrado'
    delete:
      tags:
        - Filmes
      summary: Remove um filme
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Documento removido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultadoRemocao'
        '400':
          $ref: '#/components/responses/RequisicaoInvalida'
        '404':
          $ref: '#/components/responses/NaoEncontrado'
  /api/filmes/{id}/atores:
    post:
      tags:
        - Filmes
      summary: Acrescenta um ator ao elenco com $push
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nome
                - personagem
              properties:
                nome:
                  type: string
                  example: Casey Affleck
                personagem:
                  type: string
                  example: Tom
      responses:
        '200':
          description: Filme já atualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Filme'
        '400':
          $ref: '#/components/responses/RequisicaoInvalida'
        '404':
          $ref: '#/components/responses/NaoEncontrado'
  /api/filmes/{id}/generos:
    post:
      tags:
        - Filmes
      summary: Acrescenta um gênero sem duplicar, com $addToSet
      description: Gêneros formam um conjunto — repetir não faria sentido, por isso `$addToSet` e não `$push`.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - genero
              properties:
                genero:
                  type: string
                  example: Suspense
      responses:
        '200':
          description: Filme já atualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Filme'
        '400':
          $ref: '#/components/responses/RequisicaoInvalida'
        '404':
          $ref: '#/components/responses/NaoEncontrado'
...